review fixes: scanf placeholder, sigreturn cap, tarfs linkname bound
- scanf.c: drop dead `&& 1` placeholder from scanset range condition;
  the comment had promised an endpoint check that was never wired up.
  Behavior unchanged — current overlapping-range merge is idempotent.
- signal.c: tighten sigreturn sepc/sp validation to USER_STACK_TOP so
  it agrees with trap.c's check_user_return and proc.c's clone-frame
  check. KVMEM_OFFSET was permissive enough to let a sigframe survive
  validation only to be killed by the trap-return check immediately
  afterwards.
- devfs.c: drop redundant cli/sti pair inside the TIOCSPGRP proc-list
  walk. console_ioctl's outer scope already disables interrupts, so
  the inner save read SIE=0 and the restore was a no-op.
- sh.c: emit "line too long, truncated" only after the drain loop
  confirms additional bytes arrived. Exact MAXLINE-1 chars followed by
  EOF (no newline) no longer triggers a false-positive warning.
- tarfs.c: bound hardlink h->name / h->linkname scans to the actual
  field width (post strip_prefix offset) and refuse entries with
  unterminated fields. The 254-byte bound from T3.5's hardlink add
  could over-read up to 150 bytes into adjacent ustar header fields.
  Mirrors the discipline tarfs_alloc_symlink applies (T3.6).
- proc.c: replace misleading "fall through with whatever we have"
  comment in alloc_proc with an accurate termination argument
  (live_count + 1 iterations).
- selftest.c: widen fileseek pos/fsize locals to int64_t to match
  T3.27's return-type widening; compare against ip->size as uint64_t.
- printf.c: widen tmpnam's num[12] decimal buffer to num[21] for
  full uint64_t coverage.

+++ kernel/selftest.c
@@ -464,12 +464,12 @@ static void test_file_read(void) {
     st_check(f->off == 3, "file_read: offset advanced to 3");
 
     // SEEK_CUR+0: query position.
-    int pos = fileseek(f, 0, SEEK_CUR);
+    int64_t pos = fileseek(f, 0, SEEK_CUR);
     st_check(pos == 3, "file_read: SEEK_CUR+0 == 3");
 
     // SEEK_END+0: position equals file size.
-    int fsize = fileseek(f, 0, SEEK_END);
-    st_check(fsize >= 0 && (unsigned int)fsize == ip->size,
+    int64_t fsize = fileseek(f, 0, SEEK_END);
+    st_check(fsize >= 0 && (uint64_t)fsize == ip->size,
              "file_read: SEEK_END+0 equals ip->size");
 
     // Read at EOF returns 0.
